move UndoEngine.UndoUnit to its own file
